Far more things than what the branch was supposed to contain. And I am still working on the FWLS part

+++ recs/content_based_recommender.py
@@ -1,27 +1,31 @@
 from recs.base_recommender import base_recommender
 from analytics.models import Rating
-from recommender.models import MovieDescriptions, Similarity
+from recommender.models import MovieDescriptions
 
 from gensim import models, corpora, similarities, matutils
-from scipy import spatial
 
 
 class ContentBasedRecs(base_recommender):
+
+    def __init__(self):
+        self.lda_path = './../lda/'
+
     def recommend_items(self, user_id, num=6):
+
         ratings = Rating.objects.filter(user_id=user_id)
 
-        lda = models.ldamodel.LdaModel.load('./lda/model.lda')
+        lda = models.ldamodel.LdaModel.load(self.lda_path + 'model.lda')
 
-        dictionary = corpora.Dictionary.load('./lda/dict.lda')
+        dictionary = corpora.Dictionary.load(self.lda_path + 'dict.lda')
 
-        corpus = corpora.MmCorpus('./lda/corpus.mm')
+        corpus = corpora.MmCorpus(self.lda_path + 'corpus.mm')
 
         content_sims = dict()
         for rating in ratings:
 
             md = MovieDescriptions.objects.filter(imdb_id=rating.movie_id).first()
             if md is not None:
-                index = similarities.MatrixSimilarity.load('./lda/index.lda')
+                index = similarities.MatrixSimilarity.load(lda_path + 'index.lda')
 
                 lda_vector = lda[corpus[int(md.lda_vector)]]
                 sims = index[lda_vector]
@@ -41,8 +45,8 @@ class ContentBasedRecs(base_recommender):
         ratings = Rating.objects.filter(user_id=user_id)
         rated_movies = {r['movie_id']: r['rating'] for r in ratings.values()}
 
-        lda = models.ldamodel.LdaModel.load('./lda/model.lda')
-        corpus = corpora.MmCorpus('./lda/corpus.mm')
+        lda = models.ldamodel.LdaModel.load(self.lda_path + 'model.lda')
+        corpus = corpora.MmCorpus(self.lda_path + 'corpus.mm')
 
         md = MovieDescriptions.objects.filter(imdb_id=item_id).first()
         rated_movies_desc = MovieDescriptions.objects.filter(imdb_id__in=rated_movies.keys())
@@ -50,12 +54,14 @@ class ContentBasedRecs(base_recommender):
         if md is None:
             return 0
 
+        if rated_movies_desc is None:
+            return 0
         if rated_movies_desc.count() == 0:
             return 0
 
         top = 0.0
         bottom = 0.0
-        sim_items = 0
+
         for rm in rated_movies_desc:
             lda_vector = lda[corpus[int(md.lda_vector)]]
             lda_vector_sim = lda[corpus[int(rm.lda_vector)]]
